Add switch view widgets
New widget and widget wrapper to be used with multiple views hosted in one and
same parent container. The idea is similar to the tab view components, however
instead of externally controlled visibility, each subcomponent can continue to use
its internal setVisibility method to control its visibility. This means that the
parent does not care which component is currently visible. One can listen for the SHOW
event to handle visibility switches of the children.